requirements.txt
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
colorama==0.4.4
colorclass==2.2.0
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
dnspython==1.16.0
docopt==0.6.2
eventlet==0.30.2
Fiona @ file:///C:/Users/user/pipwin/Fiona-1.8.18-cp36-cp36m-win_amd64.whl
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==4.3.1
GDAL @ file:///C:/Users/user/pipwin/GDAL-3.1.4-cp36-cp36m-win_amd64.whl
geopandas @ file:///C:/Users/user/pipwin/geopandas-0.9.0-py3-none-any.whl
greenlet==1.0.0
idna==2.10
importlib-metadata==3.7.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
Js2Py==0.70
jsonpickle==2.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
munch==2.5.0
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysql-connector-python-rf==2.2.2
mysql-utils==0.0.2
mysqlclient==2.0.3
networkx==2.5
numpy @ file:///C:/Users/user/pipwin/numpy-1.19.5%2Bvanilla-cp36-cp36m-win_amd64.whl
packaging==20.9
pandas==1.1.5
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
pip-upgrader==1.4.15
pipwin==0.5.1
prompt-toolkit==3.0.16
protobuf==3.15.5
Pygments==2.8.1
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pyproj==3.0.1
pySmartDL==1.3.4
python-dateutil==2.8.1
python-decouple==3.4
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
pyvis==0.1.9
requests==2.25.1
Rtree==0.9.7
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
tabulate==0.8.9
terminaltables==3.1.0
traitlets==4.3.3
transitions==0.8.7
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
